requirements.txt
streamlit
pandas
requests
beautifulsoup4
pdfplumber
langchain
langchain-community
langchain-openai
python-dotenv
homebrew